Update coob.py
REmoved the face transformation from 'accumulate_faces' and threw it into 'sort_faces' so I could accumulate and remove faces at the start and it would only have to run once, and then I transformed the points in the 'sort_faces'. I also made the black outline of the cube be rendered with sorted_faces so instead of having (12 edges/cube * 27 cubes)=328 edges, it no w only puts the line around (4 edges/face * 54 faces) = 216 edges which isn't as optimized but it still removes the faces in the middle

I tried out a few things with proper distance formulas to cut the number of rendered faces in half (we only need to see 3 sides of the cube) but nothing works right now. 

OH also apparently having the clock tick (60) and then (120) makes it far slower than it should be, i removed the 120 and the renderer runs at max speed.